Fixed the projection matrix so now viewable range is 0.1 to 1000 instead of 0.1 to 1.0.
Made DirectX::Blob a little more robust and more like a smart pointer.

Changed alignas to __declspec(align()) to work with VS 2013.

added an ID3D11Debug interface member only in debug mode.

Took out the model overloaded constructor in favor of a Init function, this may change back.

LoadShaderFile now takes a wstring instead of string
LoadShaderFile now returns a DirectX::Blob instead of filling in a reference to one

+++ Direct3D 11/Model.cpp
@@ -57,7 +57,8 @@ void Model_Position_Texcoord::CreatePlane(const Direct3D & D3D)
 
 	// Create the vertex buffer using the vector of vertices
 	Vertex *pVertices = vertices.data();
-	vertex_buffer = D3D.CreateVertexBuffer(vertices.size(), sizeof(Vertex),
+	vertex_buffer = D3D.CreateVertexBuffer(static_cast<int>(vertices.size()), 
+											static_cast<int>(sizeof(Vertex)),
 		reinterpret_cast<void**>(&pVertices));
 
 	// The vertices are in the correct order to be rendered, so we can just use
@@ -71,8 +72,8 @@ void Model_Position_Texcoord::CreatePlane(const Direct3D & D3D)
 
 }
 
-UINT Model_Position_Texcoord::GetVertexCount( ) const
-{
+size_t Model_Position_Texcoord::GetVertexCount( ) const
+{		
 	return vertices.size();
 }
